requirements.txt
fastapi[standard]
uvicorn

PyMySQL
peewee

python-dotenv